Progress on CategoryPage
Refactor UI to introduce tabs for Categories and SubCategories
Update SubCategory model: replace Guid Category member with CategoryId
Recreate database migrations to match model changes
Add SubCategoryWithCategoryInfoDto to display related Category info correctly in the SubCategories tab
Implement SubCategories UI, including SubCategoryAddPage.razor and SubCategoryEditPage.razor
Fix SubCategoryServerService.Update method issue
Minor improvements in CategoryAddPage, CategoryEditPage, and the Categories tab on CategoryPage
Improve CategoryServerService and CategoryController to prevent Category deletion if there are SubCategories bound to it

+++ C3P1.Net/C3P1.Net/Services/Apps/BankBook/CategoryServerService.cs
@@ -33,19 +33,22 @@ namespace C3P1.Net.Services.Apps.BankBook
 
         public async Task<bool> DeleteCategoryAsync(Guid userId, Guid categoryId)
         {
-            // find category
+            // find category and verify ownership
             var category = await _context.Categories
                 .FirstOrDefaultAsync(x => x.Id == categoryId && x.UserId == userId);
+            var subcategioriesBound = await _context.SubCategories
+                .AsNoTracking()
+                .AnyAsync(x => x.CategoryId == categoryId && x.UserId == userId);
 
-            // delete if found
-            if (category is not null)
+            // delete if found and no subcategories bound to it
+            if (category is not null && subcategioriesBound == false)
             {
                 _context.Categories.Remove(category);
                 int recorded = await _context.SaveChangesAsync();
                 return (recorded == 1);
             }
-
-            return false;
+            else
+                return false; // not found or subcategories bound
         }
 
         public async Task<List<Category>> GetCategoriesAsync(Guid userId)
